feat: rework where common functions are kept
Writing lots of TypeScript-available functions in Go was getting tedious, so this switches out to having a sidecar TypeScript file which is loaded for all the globals.

It's a little annoying, as I can't provide the types from the `implementations.ts` in the `index.d.ts` file automatically, so they're copied over to make writing the story files a little easier.

This also implements doing our own song scanning.

This prevents a confusing situation where, if you were entitled to new songs being unlocked, running the program a second time (without scanning) would mark a story as inactive, because the songs weren't in the songcache (no scan yet) *and* weren't in the hidden songs list (they weren't hidden). Now it's all managed by us, we have the added benefit of being able to determine what songs are unlocked but not yet scanned/visible.