feat: do our own song scanning
this prevents a confusing situation where, if you were entitled to new songs being unlocked, running the program a second time (without scanning) would mark a story as inactive, because the songs weren't in the songcache (no scan yet) *and* weren't in the hidden songs list (they weren't hidden). Now it's all managed by us, we have the added benefit of being able to determine what songs are unlocked but not yet scanned/visible.